Latest changes requested on PR
